moved route components into routes folder
