+++ app/src/routes/App/App.tsx
@@ -6,11 +6,11 @@ import { IconButton } from "@material-ui/core";
 import CloseIcon from "@material-ui/icons/Close";
 import FileCopyIcon from "@material-ui/icons/FileCopy";
 import { Helmet } from "react-helmet-async";
-import { Header } from "../Header";
+import { Header } from "../../components/Header";
 import { Switch, Route } from "react-router-dom";
 import { Home } from "../Home";
 import { PriceCompare } from "../PriceCompare";
-import { Faq } from "../Faq";
+import { Faq } from "../../components/Faq";
 import { makeStyles } from "@material-ui/core/styles";
 import { Graph } from "../Graph";
 import { useMediaQueryContext } from "@src/context/MediaQueryProvider";
